requirements.txt
openai
prometheus-api-client
python-dotenv
